extract the gravatar function into it's own module
